Вывод всех карт + небольшая оптимизация

+++ service/tag-service.js
@@ -48,7 +48,19 @@ module.exports = new class TagService {
         return tags;
     }
 
-    async deleteBindTagToMap(id_map, id_tag) {
+    async deleteBindTagToMap(id_user, id_map, id_tag) {
+        // Проверка, принадлежит ли карта пользователю
+        const map = await Map.findOne({
+            where: {
+                id: id_map,
+                id_creator: id_user
+            }
+        });
+
+        if (!map) {
+            throw new ApiError('Отказано в доступе к карте или карта не найдена', 403);
+        }
+
         await MapHasTag.destroy({
             where: {
                 id_map: id_map,
@@ -70,6 +82,45 @@ module.exports = new class TagService {
                     id: id_tag
                 }
             });
+        }
+    }
+
+    async deleteAllBindTagToMap(map) {
+        const { id } = map;
+    
+        // Получение всех тегов, связанных с картой
+        const tagsToDelete = await MapHasTag.findAll({
+            where: {
+                id_map: id,
+            }
+        });
+    
+        // Удаление всех привязок тегов к карте
+        await MapHasTag.destroy({
+            where: {
+                id_map: id,
+            }
+        });
+    
+        // Проверка и удаление тегов, если они не привязаны к другим картам
+        for (const tag of tagsToDelete) {
+            const tagId = tag.id_tag;
+    
+            // Проверка, используется ли тег в других картах
+            const bindedTag = await MapHasTag.findOne({
+                where: {
+                    id_tag: tagId,
+                }
+            });
+    
+            // Если тег не используется в других картах, удаляем его
+            if (!bindedTag) {
+                await Tag.destroy({
+                    where: {
+                        id: tagId,
+                    }
+                });
+            }
         }
     }
 }
